Code changed to include COMMON stocks ONLY: SHRCD=10 or 11
Formerly, my code included all CRSP stocks, not just common stocks only
as in Bali et al.
This was the primary reason that my table replications differed in
sample sizes from those of the book.

Now, only common stocks (with stocknames/msenames.SHRCD=10 or 11) are
used.
As a result, the number of firms in the table are close enough now.
However, they differ substantially in values - need to find out why.
Maybe this is because of winsorizing issues - should see the median
values.
If the median values are close enough but all others differ much, then
it would be due to extreme values, which can be alleviated by
winsorizing.