RRLOOP2: Note that "date1" is "today"
Runs a regression whose window "starts" at "date1 (which is today)".
This setting is intuitively LESS appealing than RRLOOP whose "today" is
"date2",
because one gathers the "past data" and runs regression "today".

However, if "date1" or "today" is the start of the moving window, then
it is like
as if one "foretells" the future data and runs a regression with those
"foretold" data.

Anyway, it yields technically the same result as RRLOOP does.
Hence, effectively, it's just a matter of taste.