Hard-coded upto 6M (or 6D, 6Y; upto 6 periods ahead, in fact) persistence (or cross-sectional correlation).
Hard-coded way doesn't make much sense at the first glance,
but this is the way provided by SAS in fact.

There might be a better way to do this, but this is enough for practical
purposes,
as empirical papers usually calculate this no more than 6 periods ahead.

Anyway, this could be improved.