Do not change the hard coded variable, EVENTVARS unless I don't want to see an error.
There must be a reason that EVENTVARS is hard-coded unlike SEVARS or
SFVARS.
If I let EVENTVARS to be SEVARS, it will cause a problem.
This can be checked by
PROC SORT NODUPKEY (try also NODUPLICATES/NODUP in place of NODUPKEY).

EVENTVARS seem to be those who rarely changes, and changes only if when
there's an event. (So nothing like DATE.)
--> Seems plausible guess, but exclusion of CUSIP, PERMNO, PERMCO is
unclear. Maybe this is just a matter of a taste as TICKER and NCUSIP is
included already.