SAS Enterprise Guide - Project based work.
Bad because the project's codes are not shown here.
May have to save the underlying programs separately, explicitly as files
to make the "project"
just a "wrapper" of base SAS codes.

In this way, I won't be able to change the underlying programs' names
directly,
but controlling versions and observing modifications directly here in
GitHub is valuable, which should not be missed.